refactorización de código
