+++ controllers/charola.controller.js
@@ -1,3 +1,35 @@
 const Charola = require("../models/charola.model.js");
 
-exports.registrarCharola = async (req, res) => {};
+/**
+ * @description Consultar charola ontiene todos los datos correspondientes de una charola con su ID.
+ * @param {*} req - Solicitud HTTP que contiene la ID de la charola.
+ * @param {*} res - Respuesta HTTP que se usa para enviar el resultado.
+ * @returns {JSON} Código de respuesta y token de sesión
+ */
+
+const consultarCharola = async (req, res) => {
+  const { id } = req.params;
+  try {
+    const charola = await Charola.getCharola(parseInt(id));
+
+    if (charola.error) {
+      return res.status(404).json({ error: charola.error });
+    }
+
+    res.status(200).json({
+      data: charola
+    });
+
+    console.log("Se pudo conectar el front con el back :)");
+  } catch (error) {
+    res.status(500).json({ error: error.message });
+  }
+};
+
+const registrarCharola = async (req, res) => {};
+
+module.exports = {
+  consultarCharola,
+  registrarCharola
+};
+
